cleanup: split stuff
